🚚 move/rename : AlarmCalculatorView 기존 위치로 그룹 및 파일 이동 & storyboard 이름 변경
