+++ Inception/Inception/Info.plist
@@ -16,7 +16,7 @@
 					<key>UISceneDelegateClassName</key>
 					<string>$(PRODUCT_MODULE_NAME).SceneDelegate</string>
 					<key>UISceneStoryboardFile</key>
-					<string>AlarmCalculatorViewController</string>
+					<string>AlarmCalculator</string>
 				</dict>
 			</array>
 		</dict>
